+++ react/HomePage.js
@@ -13,10 +13,8 @@ export default class HomePage extends Component {
 
   render() {
     return (
-      <div className="w-100 h-100">
-        <ExtensionPoint id="carousel" />
-        <ExtensionPoint id="shelf" />
-        <ExtensionPoint id="custom" />
+      <div className="vtex-dreamstore__container w-100 h-100">
+        <ExtensionPoint id="sections" />
       </div>
     )
   }