+++ react/components/Header.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { injectIntl, intlShape } from 'react-intl'
+import { FormattedMessage, injectIntl, intlShape } from 'react-intl'
 import Input from '@vtex/styleguide/lib/Input'
 import Button from '@vtex/styleguide/lib/Button'
 import SearchBar from 'vtex.storecomponents/SearchBar'
@@ -9,11 +9,14 @@ import { ExtensionPoint } from 'render'
 
 import SearchIcon from '../images/SearchIcon'
 
+export const TOAST_TIMEOUT = 3000
+
 class Header extends Component {
   constructor(props) {
     super(props)
     this.state = {
       searchValue: '',
+      isAddToCart: false,
     }
   }
 
@@ -28,14 +31,23 @@ class Header extends Component {
     this.setState({ searchValue: value })
   }
 
+  componentDidMount() {
+    document.addEventListener('item:add', () => {
+      this.setState({ isAddToCart: !this.state.isAddToCart })
+      window.setTimeout(() => {
+        this.setState({ isAddToCart: !this.state.isAddToCart })
+      }, TOAST_TIMEOUT)
+    })
+  }
+
   handleSearch = () => location.assign(`/${this.state.searchValue}/s`)
 
   render() {
     const { account } = global.__RUNTIME__
     const { name } = this.props
-    const { searchValue } = this.state
+    const { searchValue, isAddToCart } = this.state
     return (
-      <div>
+      <div className="relative fixed z-2 w-100 shadow-5">
         <div className="z-2 items-center w-100 top-0 bg-white tl">
           <ExtensionPoint id="menu-link" />
         </div>
@@ -61,7 +73,17 @@ class Header extends Component {
               <ExtensionPoint id="minicart" />
             </div>
           </div>
+          <div className="absolute top-3 right-1">
+            <ExtensionPoint id="minicart" />
+          </div>
         </div>
+        {isAddToCart && (
+          <div className="pa2 absolute flex justify-center w-100">
+            <Alert type="success" autoClose={TOAST_TIMEOUT}>
+              <FormattedMessage id="dreamstore.buy-success" />
+            </Alert>
+          </div>
+        )}
       </div>
     )
   }